refactor: UserOutcome / classifyOutcomeForUser をドメイン層からアプリケーション層に移動 (#596)
Match Aggregate の不変条件ではなく表示・集計用途の変換ロジックであるため、
server/domain/models/match/ から server/application/match/ に移動。
テストも合わせて移動し、利用箇所のインポートパスを更新。
