+++ server/domain/models/match/match.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, test } from "vitest";
 import { circleSessionId, matchId, userId } from "@/server/domain/common/ids";
 import {
-  classifyOutcomeForUser,
   createMatch,
   deleteMatch,
   hasDifferentPlayers,
@@ -171,39 +170,5 @@ describe("Match ドメイン", () => {
 
   test("hasDifferentPlayers は同一プレイヤーで false を返す", () => {
     expect(hasDifferentPlayers(userId("user-1"), userId("user-1"))).toBe(false);
-  });
-
-  describe("classifyOutcomeForUser", () => {
-    test("P1_WIN でプレイヤー1の場合 win を返す", () => {
-      expect(classifyOutcomeForUser("P1_WIN", true)).toBe("win");
-    });
-
-    test("P1_WIN でプレイヤー2の場合 loss を返す", () => {
-      expect(classifyOutcomeForUser("P1_WIN", false)).toBe("loss");
-    });
-
-    test("P2_WIN でプレイヤー1の場合 loss を返す", () => {
-      expect(classifyOutcomeForUser("P2_WIN", true)).toBe("loss");
-    });
-
-    test("P2_WIN でプレイヤー2の場合 win を返す", () => {
-      expect(classifyOutcomeForUser("P2_WIN", false)).toBe("win");
-    });
-
-    test("DRAW でプレイヤー1の場合 draw を返す", () => {
-      expect(classifyOutcomeForUser("DRAW", true)).toBe("draw");
-    });
-
-    test("DRAW でプレイヤー2の場合 draw を返す", () => {
-      expect(classifyOutcomeForUser("DRAW", false)).toBe("draw");
-    });
-
-    test("UNKNOWN でプレイヤー1の場合 null を返す", () => {
-      expect(classifyOutcomeForUser("UNKNOWN", true)).toBeNull();
-    });
-
-    test("UNKNOWN でプレイヤー2の場合 null を返す", () => {
-      expect(classifyOutcomeForUser("UNKNOWN", false)).toBeNull();
-    });
   });
 });
