+++ server/domain/models/match/match.ts
@@ -101,25 +101,3 @@ export const deleteMatch = (match: Match, deletedAt?: Date): Match => ({
   ...match,
   deletedAt: deletedAt ?? new Date(),
 });
-
-export type UserOutcome = "win" | "loss" | "draw";
-
-export const classifyOutcomeForUser = (
-  outcome: MatchOutcome,
-  isPlayer1: boolean,
-): UserOutcome | null => {
-  switch (outcome) {
-    case "UNKNOWN":
-      return null;
-    case "DRAW":
-      return "draw";
-    case "P1_WIN":
-      return isPlayer1 ? "win" : "loss";
-    case "P2_WIN":
-      return isPlayer1 ? "loss" : "win";
-    default: {
-      const _exhaustive: never = outcome;
-      return _exhaustive;
-    }
-  }
-};
